backend/requirements.txt
flask>=3.0
flask-cors>=4.0
pandas>=2.1
numpy>=1.26
scikit-learn>=1.4
joblib>=1.3
mysql-connector-python>=8.3
geojson>=3.1
matplotlib>=3.8
seaborn>=0.13
